Parallel mode; notes about using open z-wave 1.6
Hass 0.113.0 added modes for scripts.  I've added "mode: parallel" so switches can be updated a little faster.  If this overwhelms your z-wave network, change it to "queued" so the lights are updated sequentially.  

Hass 0.113.0 also added set_config_parameter functionality for open z-wave 1.6 (currently still in beta).  Until I figure out how to auto-detect if a user is running 1.6 or the default 1.4 z-wave integration, the service will have to be manually changed.  For now I'm defaulting to the z-wave 1.4 parameter, since that's the default supported integration.