Complete re-write for zwavejs2mqtt container 2.2.0
Still calculates the effect for dimmers, which blocks pulse since the value is greater than 83823359.  Inputs are different than the original script, so we're handling it as a new file.  This would let somebody disable their old automations, duplicate them, and slowly translate them to the new zwave-js script, and revert back to ozw1.4 if they need to.  That might be handy for somebody.