Updated for zwavejs2mqtt container 2.4.0
The dimmer effect parameters are no longer calculated in zwavejs2mqtt container 2.4.0, and get passed individually like the other devices.  This means "pulse" will work as an effect again.  An indirect result is that the "choose" options are cleaner.

"Intensity" changed to "Brightness" in the parameter names, so I've changed the variables to match.  It's a tacky thing to do, but nobody was using the 2.2.0 version of the script yet, and it's not backwards compatible with the version of the script for Open Z-Wave 1.4 anyway.