Selectors!: The Revenge
The target selector is passing more than just the entity ID, so I'll have to figure that out later.  For now I'm switching to an entity selector.  I've also removed the domain: filter, since the combo_fan entities aren't lights.  Maybe I'll rewrite it to accept a light_entity, fan_entity, and combo_light_entity, but if more than one is populated that would need to be handled in the script somehow.  Another day, perhaps…