The Defaults Strike Back!
Brightness variables now default to 0—which is invalid in the selector—so that they're skipped if they're not being used.  Previously, if a user was setting an effect and not setting the non-effect brightness, the non-effect brightness default value would alter the configuration parameter and when the effect cleared, the LED strip wouldn't return to its original value.  Now an effect can be set or unset without modifying the underlying LED brightness.