Accepts any combo of areas, devices, and entities with mixed device types.
The script can now be called with an area, device, or entity of a device, with any combination of the three.  It will identify the entities within the area or device and assign them to their proper type so that we no longer need to define the model or call the script multiple times when dealing with different device types (i.e. dimmers vs switches).  It should still be backwards compatible with automations passing entities in a comma-separated list, so you can update to this script without tracking down each automation that's calling it.