Handle Strings for Areas and Devices, and Model Detection Clean-Up
Aeas and devices are no longer expected to be in a list format.  This should make calling the script much easier when you're not using the UI.  I've also made slight modifications to the model detection process to make it a tiny bit faster (and easier to maintain and debug).