Handles area: all and Improved Devices Selector
*** Home Assistant 2023.04 or newer is required ***

Using the new "areas()" function introduced in 2023.04 when "area: all" is passed to the script.  It will identify all areas in your home and find all compatible Inovelli devices to apply the settings to.  

service: script.inovelli_led_zwavejs
data:
  area: all
  duration: Forever
  effect: CHASE
  brightness: 8
  color: Teal

The devices selector was also updated with an improved filter so that ONLY compatible devices are offered.  I wish I could get the area and entity selectors more precisely tuned, but it doesn't seem to be possible (https://community.home-assistant.io/t/area-selector-device-does-not-limit-selector-options/565184).