Adding LED color when off and effects from Blue and Red 800 Series devices
Inovelli's Red 800 Series Z-Wave devices and their Blue Series Zigbee devices support more effects than the original Red Series devices.  I've added those effects and remapped them to in the original Red Series devices to something I think is reasonable.

We'll also support LEDcolor_off as an option now and only send it to Red 800 Series and Blue Series devices.  The original Red Series devices don't support a different color when the LED is off.  

The new devices also have a different default color set but in order to maintain a consistent user experience for those already calling "teal" I'm not remapping the values.  I'm not sure these are adding a lot of value compared to what I already had and I'm tempted to remove them and shrink the list.  It's a getting unwieldy to look at in the UI.