Fix LZW36-SN Effect Clear
Effect duration wasn't being set to 0 in the default path (clearing effects) when effect type was being set to off.  As a result, the LEDcolor settings didn't work because the devices was displaying an "off" effect with no timeout. 

This should have affected similar behavior in other devices as well and I'm not sure why it didn't.