Fix zha crash, better z2m entity detection, and Zigbee2MQTT 2.0.0 compatibility (#36)
* Performance and stability improvements in Blue Series

* Zigbee2MQTT entities are now detected with `integration_entities('mqtt')` instead of not zha.
* Blue Series LED color and brightness now runs in about one-third the time it needed previously.
* My ZHA and Z2M networks seem more stable.

* See notes on the script.

* Z2M LED Brightness Can be Cleared

Setting `LEDcolor: all clear` and `LEDcolor_off: all clear` were clearing the individual LED colors but not the brightness levels.  As a results, LED brightness levels couldn't be modified by the `LedIntensityWhenOn` and `LedIntensityWhenOff` parameters in the Z2M UI.

* ZHA and Individual LED Brightness Can be Cleared

ZHA now clears like Z2M, and individual LEDs can be cleared for both (not just the full set).

* Further fixes to LED brightness on / off logic

*** BREAKING CHANGE ***

If you're using Blue Series devices this change impacts you.  The script was previously setting LED brightness levels and LED colors by individual LED.  This overrides the parameter which controls the LED bar as a single, uniform option.  I've fixed the logic for how this is done, which means LED color and brightness levels only need one service call for each option, instead of eight (one for treating the LED bar as a single unit and one for each of the seven LEDs).  This should speed things up but the individual LEDs need to be cleared so they don't override single unit setting.  In order to do that, call the script (or set up an instance of the blueprint) with these settings:

```
  area:
    - all
  LEDnumber: All
  LEDcolor: All Clear
  LEDcolor_off: All Clear
  LEDbrightness: 7
  LEDbrightness_off: 2
```
That should set all of your LED bars to red.  Once that's done, your old automations and blueprints should work just fine.  This is a one-time issue.

* Zigbee2MQTT 2.0.0 Compatibility

The model name changed… again.

* Zigbee2MQTT Compatibility