Handling devices without model (#46)
* Handling Devices Without a Model ID

Devices without a model_id were generating `TypeError: argument of type 'NoneType' is not iterable` errors.  Thanks to @snapcase for figuring this out.  The solution I've settled on is to filter for Inovelli devices when I establish the initial list of entities.  This should speed up the script a tiny bit, I believe, but it's also much more reliable and easier to debug than filtering out bad entities.