White Series Brightness and Color
Added brightness and color configuration for White Series devices.  As far as I can tell, only the "_6" entity can be configured, which means the level of the "_1" entity light can't be reflected in the LED bar.  This all seems like a bad hack job.  Maybe there's something I'm not understanding.