Better White Series Handling
LED bar can now change color when on so it displays the dim level of the device.  This is done through a `select.*` entity and guessing at the name.  If you've changed the name of your light or select entities, this will break.  One needs to start with `light.` and end with `_light_1` while the other needs to start with `select.` and end with `led_color`.  The rest of the `object_id` needs to match.  

LED bar color when off is not supported (not possible).

LED bar brightness and brightness when off is not supported (not possible).

The only effect supported is "solid", which is done with a secondary `light.*` entity.  This name doesn't need to match the primary `light.*` entity like the LED bar color `switch.*` entity does. 

Overall, this still feels like a hack to me.